Use Traktor VOLUME attribute
On closer inspection, the previous issue with COLLECTION keys
not matching PLAYLIST was because we were building the key based on the
LOCATION.VOLUMEID which has odd capitalization issues.  The
LOCATION.VOLUME key however matched the PLAYLIST key so we can remove
the lowercasing and support case insensitive file systems again.